new endpoint and blackbox tests
